eu-registry-bot/requirements.txt
# EU Registry Bot - Dependencies

# Browser Automation
selenium>=4.15.0
webdriver-manager>=4.0.1

# Certificate Handling
cryptography>=41.0.0
pyOpenSSL>=23.3.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Scheduling
APScheduler>=3.10.4

# HTTP Requests
requests>=2.31.0
urllib3>=2.1.0

# Logging & Monitoring
colorlog>=6.8.0

# Data Handling
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: CAPTCHA Solving (uncomment if needed)
# 2captcha-python>=1.2.0
